@@ -30,73 +30,73 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.2",
-    "@angular/common": "~13.0.2",
-    "@angular/compiler": "~13.0.2",
-    "@angular/core": "~13.0.2",
-    "@angular/forms": "~13.0.2",
-    "@angular/platform-browser": "~13.0.2",
-    "@angular/platform-browser-dynamic": "~13.0.2",
-    "@angular/router": "~13.0.2",
-    "@ngneat/lib": "^3.0.0",
-    "@ngrx/effects": "^13.0.1",
-    "@ngrx/router-store": "^13.0.1",
-    "@ngrx/store": "^13.0.1",
-    "@reduxjs/toolkit": "~1.6.2",
+    "@angular/animations": "~13.1.1",
+    "@angular/common": "~13.1.1",
+    "@angular/compiler": "~13.1.1",
+    "@angular/core": "~13.1.1",
+    "@angular/forms": "~13.1.1",
+    "@angular/platform-browser": "~13.1.1",
+    "@angular/platform-browser-dynamic": "~13.1.1",
+    "@angular/router": "~13.1.1",
+    "@ngneat/lib": "^3.0.1",
+    "@ngrx/effects": "^13.0.2",
+    "@ngrx/router-store": "^13.0.2",
+    "@ngrx/store": "^13.0.2",
+    "@reduxjs/toolkit": "~1.7.1",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^13.0.0",
-    "@angular-devkit/build-angular": "^13.0.3",
+    "@angular-builders/jest": "^13.0.2",
+    "@angular-devkit/build-angular": "^13.1.2",
     "@angular-eslint/builder": "^13.0.1",
     "@angular-eslint/eslint-plugin": "^13.0.1",
     "@angular-eslint/eslint-plugin-template": "^13.0.1",
     "@angular-eslint/schematics": "^13.0.1",
     "@angular-eslint/template-parser": "^13.0.1",
-    "@angular/cli": "~13.0.3",
-    "@angular/compiler-cli": "~13.0.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.1",
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
-    "@ngrx/store-devtools": "^13.0.1",
+    "@ngrx/store-devtools": "^13.0.2",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
-    "@tailwindcss/forms": "^0.3.4",
-    "@testing-library/angular": "^10.11.1",
-    "@testing-library/jest-dom": "^5.15.1",
+    "@tailwindcss/forms": "^0.4.0",
+    "@testing-library/angular": "^11.0.2",
+    "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.3",
     "@types/node": "^12.20.37",
     "@types/node-fetch": "^2.5.12",
-    "@typescript-eslint/eslint-plugin": "5.5.0",
-    "@typescript-eslint/parser": "5.5.0",
+    "@typescript-eslint/eslint-plugin": "5.7.0",
+    "@typescript-eslint/parser": "5.7.0",
     "all-contributors-cli": "^6.20.0",
-    "angular-cli-ghpages": "1.0.0-rc.2",
+    "angular-cli-ghpages": "1.0.0",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.3.0",
+    "eslint": "^8.4.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "2.25.3",
     "eslint-plugin-jest": "^25.3.0",
     "eslint-plugin-jest-dom": "^3.9.2",
-    "eslint-plugin-jsdoc": "^37.0.3",
-    "eslint-plugin-ngrx": "^2.0.0",
+    "eslint-plugin-jsdoc": "^37.2.2",
+    "eslint-plugin-ngrx": "^2.0.3",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-rxjs": "^4.0.3",
-    "eslint-plugin-testing-library": "^5.0.0",
+    "eslint-plugin-testing-library": "^5.0.1",
     "git-cz": "^4.8.0",
     "husky": "^7.0.4",
-    "jest": "^27.4.2",
+    "jest": "^27.4.5",
     "jest-preset-angular": "^11.0.1",
     "jsonc-parser": "^3.0.0",
     "lint-staged": "^12.1.2",
-    "msw": "^0.35.0",
-    "ng-packagr": "^13.0.8",
+    "msw": "^0.36.3",
+    "ng-packagr": "^13.1.2",
     "node-fetch": "^2.6.6",
-    "postcss": "^8.4.4",
-    "prettier": "^2.5.0",
+    "postcss": "^8.4.5",
+    "prettier": "^2.5.1",
     "prettier-plugin-organize-attributes": "0.0.4",
     "protractor": "~7.0.0",
     "semantic-release": "^18.0.1",
