@@ -1,4 +1,4 @@
-function is(x: any, y: any): boolean {
+function is(x: unknown, y: unknown) {
   if (x === y) {
     return x !== 0 || y !== 0 || 1 / x === 1 / y;
   } else {
