@@ -1,7 +1,7 @@
 import { worker } from '../mocks/browser';
 
 // add msw mock on dev environment && e2e tests
-worker.start();
+worker.start({ onUnhandledRequest: 'bypass' });
 
 export const environment = {
   production: false,
